Add initial plugin for exposing build status
This uses the node-ipc library to expose two methods: `getAllErrors`, which does
a slow full check of all files in the project (just like tsc does during normal
typecheck), and `getErrorsForFiles`, which does a faster check of a specific set
of files. Later, the CLI tool will call `getErrorsForFiles` first for faster
feedback, then `getAllErrors` as a follow-up.