Refactor plugin to work when multiple TS projects are loaded (#3)
This is pretty much a rewrite of the plugin to better handle situations where a
single editor is working with files from multiple TS projects. Previously, the
init and create functions were being called once per project, so when loading
the second project we'd register a new IPC listener and clobber the old one, so
any file requests received from the wrong project would get a "file not found"
error and crash the TSStatusClient.

After this rewrite, we bookkeep the open projects and register a single IPC
listener that tries each project and combines the results. I also refactored it
into two classes and smaller functions and changed the logging system to just
append to a hard-coded file.

This plugin is still non-ideal for multi-project use cases because if you happen
to have the projects open in different editors, they'll run in different TS
server processes and the second listener will clobber the first. There doesn't
seem to be an obvious way to configure node-ipc to have multiple servers and get
a response from each of them, but I'm sure there's a way to make it work.

Test Plan:
Enable logging and confirm
Tweak the code so that the "not found in any project" message always appears and
make sure it that it looks reasonable.
Try using `bin/dev check` in the aurelia project with PyCharm and with VSCode
and ensure that both now work.